+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.0.0-alpha.1](https://github.com/ecomclub/storefront/compare/v0.9.0...v1.0.0-alpha.1) (2019-08-13)
+
+
+### refactor
+
+* **template:** using ecomplus/storefront-template as base ([0263022](https://github.com/ecomclub/storefront/commit/0263022))
+
+
+### BREAKING CHANGES
+
+* **template:** code structure totally changed
+
+
+
 ## [0.9.0](https://github.com/ecomclub/storefront/compare/v0.8.0...v0.9.0) (2019-07-26)